dev: Use the development environment in ./devel/generate-version
…if the environment is available.  It won't be in CI, but that's ok
because CI runs generate-version with GNU coreutils ambiently available.

This lets us add GNU coreutils to our dev env so that ./devel/build
works on macOS.

For good measure, also move the setting of PATH in ./devel/build to the
first line of main() so that everything after runs with the dev env
available.  Only one of these changes would be sufficient to address the
issue, but the latter change means we may also avoid other issues in the
future.