devel/setup: install anaconda-client from conda-forge
Instead of installing anaconda-client separately from the default
channel, install with other packages from conda-forge.

Previously we were installing from default because the version
on conda-forge was really old. Since then, the latest version of
anaconda-client has been uploaded to conda-forge.

The version on conda-forge is not limited to Python <= 3.10, so we
will not run into issues with Python 3.11.