dev: Switch away from the now-deprecated provision-with-micromamba
It was deprecated last week in favor of a new setup-micromamba action¹,
but that doesn't put `micromamba` on PATH without extra faffing about.

Since we use the action merely to obtain a `micromamba` executable on the
PATH—shell init, environment creation, etc. is all unnecessary for our
purposes—switch to the simpler installer shell script.²

¹ <https://github.com/mamba-org/provision-with-micromamba/releases/tag/v16>
² <https://mamba.readthedocs.io/en/latest/installation.html#install-script>