Add envdir program
This provides an `envdir` program to our Conda runtime since our other
runtimes also provide this useful program and we've felt its absence.¹

Since there's no existing Conda packaging of the Python envdir
distribution² we use elsewhere or any other `envdir` program AFAICT³,
the particular program provided is a port I wrote long ago⁴ and still
use daily on my own machines.  It's vendored into this repository and
built into the Conda packages we're already producing here.  This is a
quick and low-overhead solution, and we may want to switch to a
packaging of the Python envdir distribution later (particularly if we
want to use the Python envdir module that distribution also provides).

Resolves <https://github.com/nextstrain/conda-base/issues/35>.

¹ <https://github.com/nextstrain/.github/pull/44#discussion_r1214650282>
² <https://pypi.org/project/envdir/>

³ I checked for both daemontools, which provides the original, and s6,
  which provides another port, as well as a general search for "envdir".

⁴ <https://github.com/tsibley/envdir>