Pin `awscli` to resolve `devel/build` hanging
`devel/build` was hanging and failing for ubuntu-22.04 and macos-13 in
the CI workflow, but succeeded for macos-14. The downgrade of `awscli`
in the successful macos-14 summary looked suspicious, so I decided to
pin `awscli >=2` to see if it resolved the hanging issue and it did!¹

Still not sure _why_ this worked, but I'm guessing there was some weird
dependency resolution during `boa build`.

¹ <https://github.com/nextstrain/conda-base/issues/105#issuecomment-2810864347>