Remove ebola repo check
This repo has been reworked to be compatible with the newer
pathogen-repo-ci and thus no longer works with pathogen-repo-ci-v0.

I chose not to check the new version because it is still in progress.
The only check would be on an unstable¹ and long-running² ingest
workflow, which doesn't seem useful to include here as-is.

¹ https://github.com/nextstrain/ebola/issues/20
² 4 minutes - https://github.com/nextstrain/ebola/pull/15#issuecomment-3286017943